+++ src/vumi2/applications/turn_channels_api/turn_channels_api.py
@@ -123,6 +123,7 @@ class TurnChannelsApi(BaseWorker):
     config: TurnChannelsApiConfig
 
     async def setup(self) -> None:
+        await super().setup()
         message_cache_class = class_from_string(self.config.message_cache_class)
         self.message_cache = message_cache_class(self.config.message_cache_config)
         self.connector = await self.setup_receive_inbound_connector(