+++ src/vumi2/workers.py
@@ -1,4 +1,5 @@
 import importlib.metadata
+from functools import wraps
 from logging import getLogger
 from typing import TypedDict, TypeVar
 
@@ -11,7 +12,8 @@ from hypercorn.trio import serve as hc_serve
 from quart_trio import QuartTrio
 from trio.abc import AsyncResource
 
-from vumi2.config import BaseConfig
+from vumi2.class_helpers import class_from_string
+from vumi2.config import BaseConfig, structure_config
 from vumi2.connectors import (
     ConnectorCollection,
     EventCallbackType,
@@ -77,6 +79,12 @@ class BaseWorker(AsyncResource):
         self.healthchecks = {"amqp": self._amqp_healthcheck}
         if config.http_bind is not None:
             self._setup_http(config.http_bind)
+        self.middlewares = []
+        for middleware_config in self.config.middlewares:
+            middleware_class = class_from_string(middleware_config["class_path"])
+            config = structure_config(middleware_config, middleware_class)
+            middleware = middleware_class(config)
+            self.middlewares.append(middleware)
 
     def _setup_sentry(self):
         if not self.config.sentry_dsn:
@@ -126,7 +134,43 @@ class BaseWorker(AsyncResource):
         self._closed.set()
 
     async def setup(self):
-        pass
+        for middleware in self.middlewares:
+            await middleware.setup()
+
+    def middleware_outbound_handler(
+        self, connector_name, middlewares, outbound_handler
+    ):
+        middlewares = [m for m in middlewares if m.outbound_enabled(connector_name)]
+
+        @wraps(outbound_handler)
+        async def _outbound_handler(msg):
+            for middleware in middlewares:
+                msg = await middleware.handle_outbound(msg, connector_name)
+            return await outbound_handler(msg)
+
+        return _outbound_handler
+
+    def middleware_inbound_handler(self, connector_name, middlewares, inbound_handler):
+        middlewares = [m for m in middlewares if m.inbound_enabled(connector_name)]
+
+        @wraps(inbound_handler)
+        async def _inbound_handler(msg):
+            for middleware in middlewares:
+                msg = await middleware.handle_inbound(msg, connector_name)
+            return await inbound_handler(msg)
+
+        return _inbound_handler
+
+    def middleware_event_handler(self, connector_name, middlewares, event_handler):
+        middlewares = [m for m in middlewares if m.event_enabled(connector_name)]
+
+        @wraps(event_handler)
+        async def _event_handler(msg):
+            for middleware in middlewares:
+                msg = await middleware.handle_event(msg, connector_name)
+            return await event_handler(msg)
+
+        return _event_handler
 
     async def start_consuming(self):
         await self._connectors.start_consuming()
@@ -142,6 +186,12 @@ class BaseWorker(AsyncResource):
                 "Attempt to add duplicate receive inbound connector with name"
                 f" {connector_name}"
             )
+        _inbound_handler = self.middleware_inbound_handler(
+            connector_name, self.middlewares, inbound_handler
+        )
+        _event_handler = self.middleware_event_handler(
+            connector_name, self.middlewares, event_handler
+        )
         connector = ReceiveInboundConnector(
             self.nursery,
             self.connection,
@@ -150,7 +200,7 @@ class BaseWorker(AsyncResource):
         )
         self._connectors.add(connector)
         await connector.setup(
-            inbound_handler=inbound_handler, event_handler=event_handler
+            inbound_handler=_inbound_handler, event_handler=_event_handler
         )
         self.receive_inbound_connectors[connector_name] = connector
         return connector
@@ -163,6 +213,10 @@ class BaseWorker(AsyncResource):
                 "Attempt to add duplicate receive outbound connector with name"
                 f" {connector_name}"
             )
+        _outbound_handler = self.middleware_outbound_handler(
+            connector_name, self.middlewares, outbound_handler
+        )
+
         connector = ReceiveOutboundConnector(
             self.nursery,
             self.connection,
@@ -170,6 +224,6 @@ class BaseWorker(AsyncResource):
             self.config.worker_concurrency,
         )
         self._connectors.add(connector)
-        await connector.setup(outbound_handler=outbound_handler)
+        await connector.setup(outbound_handler=_outbound_handler)
         self.receive_outbound_connectors[connector_name] = connector
         return connector